Move genomic island queries to a Query Class.
